feat: Pet Category Api dev.

+++ src/main/java/LittlePet/UMC/SmallPet/repository/PetCategoryRepository.java
@@ -3,5 +3,5 @@ package LittlePet.UMC.SmallPet.repository;
 import LittlePet.UMC.domain.petEntity.categories.PetCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PetCategoryRepository extends JpaRepository<PetCategory, Integer> {
+public interface PetCategoryRepository extends JpaRepository<PetCategory, Long> {
 }
